+++ packages/amplify-graphql-api-construct/src/__tests__/__functional__/functions.test.ts
@@ -2,18 +2,19 @@ import * as cdk from 'aws-cdk-lib';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
 import { Match, Template } from 'aws-cdk-lib/assertions';
 import { AmplifyGraphqlApi } from '../../amplify-graphql-api';
+import { AmplifyGraphqlDefinition } from '../../amplify-graphql-definition';
 
 describe('function directive', () => {
   it('references a function by name defined elsewhere', () => {
     const stack = new cdk.Stack();
     const api = new AmplifyGraphqlApi(stack, 'TestApi', {
       apiName: 'MyApi',
-      schema: /* GraphQL */ `
+      definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
         type Query {
           repeat(message: String!): String! @function(name: "repeat")
         }
-      `,
-      authorizationConfig: {
+      `),
+      authorizationModes: {
         apiKeyConfig: { expires: cdk.Duration.days(7) },
       },
     });
@@ -108,15 +109,15 @@ describe('function directive', () => {
       () =>
         new AmplifyGraphqlApi(stack, 'TestApi', {
           apiName: 'MyApi',
-          schema: /* GraphQL */ `
+          definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
             type Query {
               repeat(message: String!): String! @function(name: "repeat")
             }
-          `,
+          `),
           functionNameMap: {
             repeat: referencedFunction,
           },
-          authorizationConfig: {
+          authorizationModes: {
             apiKeyConfig: { expires: cdk.Duration.days(7) },
           },
         }),
@@ -136,15 +137,15 @@ describe('function directive', () => {
 
     const api = new AmplifyGraphqlApi(stack, 'TestApi', {
       apiName: 'MyApi',
-      schema: /* GraphQL */ `
+      definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
         type Query {
           repeat(message: String!): String! @function(name: "repeat")
         }
-      `,
+      `),
       functionNameMap: {
         repeat: referencedFunction,
       },
-      authorizationConfig: {
+      authorizationModes: {
         apiKeyConfig: { expires: cdk.Duration.days(7) },
       },
     });
@@ -188,15 +189,15 @@ describe('function directive', () => {
 
     const api = new AmplifyGraphqlApi(stack, 'TestApi', {
       apiName: 'MyApi',
-      schema: /* GraphQL */ `
+      definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
         type Query {
           repeat(message: String!): String! @function(name: "repeat")
         }
-      `,
+      `),
       functionNameMap: {
         repeat: referencedFunction,
       },
-      authorizationConfig: {
+      authorizationModes: {
         apiKeyConfig: { expires: cdk.Duration.days(7) },
       },
     });