+++ packages/amplify-graphql-api-construct/src/index.ts
@@ -4,10 +4,8 @@ export type {
   OIDCAuthorizationConfig,
   ApiKeyAuthorizationConfig,
   LambdaAuthorizationConfig,
-  AuthorizationConfig,
-  AmplifyApiGraphqlSchema,
-  AmplifyApiSchemaPreprocessor,
-  AmplifyApiSchemaPreprocessorOutput,
+  AuthorizationModes,
+  PartialTranslationBehavior,
   AmplifyGraphqlApiProps,
   AmplifyGraphqlApiResources,
   AmplifyGraphqlApiCfnResources,
@@ -19,15 +17,16 @@ export type {
   FunctionSlotOverride,
   ConflictResolution,
   ConflictDetectionType,
-  ConflictHandlerType,
   OptimisticConflictResolutionStrategy,
   CustomConflictResolutionStrategy,
   AutomergeConflictResolutionStrategy,
   ConflictResolutionStrategyBase,
   ConflictResolutionStrategy,
-  SchemaTranslationBehavior,
+  TranslationBehavior,
+  IAmplifyGraphqlDefinition,
+  IBackendOutputStorageStrategy,
+  IBackendOutputEntry,
+  AddFunctionProps,
 } from './types';
 export { AmplifyGraphqlApi } from './amplify-graphql-api';
-
-// remove these exports when provided by cli
-export { GraphqlOutput, BackendOutputStorageStrategy, BackendOutputEntry, versionedGraphqlOutputSchema } from './graphql-output';
+export { AmplifyGraphqlDefinition } from './amplify-graphql-definition';