+++ packages/amplify-graphql-maps-to-transformer/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-maps-to-transformer@3.2.0...@aws-amplify/graphql-maps-to-transformer@3.2.1) (2023-09-20)
+
+**Note:** Version bump only for package @aws-amplify/graphql-maps-to-transformer
+
+# [3.2.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-maps-to-transformer@3.1.3...@aws-amplify/graphql-maps-to-transformer@3.2.0) (2023-09-13)
+
+### Features
+
+- enable jsii builds for the api construct ([#1840](https://github.com/aws-amplify/amplify-category-api/issues/1840)) ([f6e9aff](https://github.com/aws-amplify/amplify-category-api/commit/f6e9aff2b7e8ad620e1899d29c8cb330b4d6a30f))
+
+## [3.1.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-maps-to-transformer@3.1.2...@aws-amplify/graphql-maps-to-transformer@3.1.3) (2023-09-07)
+
+**Note:** Version bump only for package @aws-amplify/graphql-maps-to-transformer
+
 ## [3.1.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-maps-to-transformer@3.1.1...@aws-amplify/graphql-maps-to-transformer@3.1.2) (2023-08-30)
 
 **Note:** Version bump only for package @aws-amplify/graphql-maps-to-transformer