+++ packages/amplify-util-mock/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.3.0](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-util-mock@6.2.4...amplify-category-api-util-mock@6.3.0) (2023-09-20)
+
+### Features
+
+- disable amplify cfn outputs for cdk apps ([0c72d18](https://github.com/aws-amplify/amplify-category-api/commit/0c72d1822f8e5ccb3e04a0a49049a459b5fb49e6))
+
+## [6.2.4](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-util-mock@6.2.3...amplify-category-api-util-mock@6.2.4) (2023-09-13)
+
+**Note:** Version bump only for package amplify-category-api-util-mock
+
+## [6.2.3](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-util-mock@6.2.2...amplify-category-api-util-mock@6.2.3) (2023-09-07)
+
+**Note:** Version bump only for package amplify-category-api-util-mock
+
 ## [6.2.2](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-util-mock@6.2.1...amplify-category-api-util-mock@6.2.2) (2023-08-30)
 
 **Note:** Version bump only for package amplify-category-api-util-mock