+++ packages/graphql-transformers-e2e-tests/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.5.5](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-graphql-transformers-e2e-tests@8.5.4...amplify-category-api-graphql-transformers-e2e-tests@8.5.5) (2023-09-20)
+
+**Note:** Version bump only for package amplify-category-api-graphql-transformers-e2e-tests
+
+## [8.5.4](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-graphql-transformers-e2e-tests@8.5.3...amplify-category-api-graphql-transformers-e2e-tests@8.5.4) (2023-09-13)
+
+**Note:** Version bump only for package amplify-category-api-graphql-transformers-e2e-tests
+
+## [8.5.3](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-graphql-transformers-e2e-tests@8.5.2...amplify-category-api-graphql-transformers-e2e-tests@8.5.3) (2023-09-07)
+
+**Note:** Version bump only for package amplify-category-api-graphql-transformers-e2e-tests
+
 ## [8.5.2](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-graphql-transformers-e2e-tests@8.5.1...amplify-category-api-graphql-transformers-e2e-tests@8.5.2) (2023-08-30)
 
 **Note:** Version bump only for package amplify-category-api-graphql-transformers-e2e-tests